Majas darbs homeworkTwo labots uzdevums8

+++ src/main/java/homeworkTwo/Uzdevums8.java
@@ -2,7 +2,7 @@ package homeworkTwo;
 
 // 8. uzdevums
 public class Uzdevums8 {
-    public static double average (int a, int b, int c){
+    public static double average (double a, double b, double c){
         double average = (a + b + c) / 3;
         return(average);
     }
